+++ src/hotspin-core/engine/solve_heun.go
@@ -67,7 +67,7 @@ func (s *HeunSolver) Step() {
 		dy := equation[i].input[0]
 		dyMul := dy.multiplier
 
-		h := float32(dt * dyMul[0])
+		h := float64(dt * dyMul[0])
 		gpu.MAdd2Async(y.Array(), dy.Array(), 0.5*h, s.buffer[i], -0.5*h, y.Array().Stream) // corrected step
 		y.Array().Sync()
 		Pool.Recycle(&s.buffer[i])
